requirements.txt
bumpversion ==0.5.3
pymongo >=3.3.0
pytest >=3.0.5
pytest-runner >=2.11
pyyaml >=3.12
fuzzywuzzy >=0.15.0
mongoengine >=0.11.0
sramongo >=0.0.1
